Refactor user+dt runners into externalCompilerRunner
