+++ src/harness/tsconfig.json
@@ -92,8 +92,7 @@
         "projectsRunner.ts",
         "loggedIO.ts",
         "rwcRunner.ts",
-        "userRunner.ts",
-        "definitelyRunner.ts",
+        "externalCompileRunner.ts",
         "test262Runner.ts",
         "./parallel/shared.ts",
         "./parallel/host.ts",
